agrande las letras de los titulos del Footer y modifique la hoja de estilos del footer
