+++ package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "bootstrap": "^5.3.3",
     "prop-types": "^15.8.1",
-    "bootstrap-icons": "^1.11.3",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.5",
     "react-dom": "^18.3.1",
