+++ src/components/common/Footer/Footer.css
@@ -1,7 +1,25 @@
+h5 {
+  font-size: 3rem;
+  letter-spacing: 0.2rem;
+  margin-bottom: 1rem; 
+}
+
 .footer-background {
   position: relative;
 
 }
+
+.text-link {
+  color: #00652F;
+  text-decoration: none;
+}
+
+.text-link:hover {
+  color: #218838;
+  text-decoration: underline; 
+  transition: color 0.3s ease;
+}
+
 
 .footer-background::before {
   content: '';
